eval/requirements.txt
pytest
azure-ai-evaluation
